refactor: only 1 controller to get mapping JSON or HTML
